@@ -1,15 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <GenerateProgramFile>false</GenerateProgramFile>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="StackExchange.Exceptional.AspNetCore" Version="2.2.32" />
-    <PackageReference Include="Hangfire.Console" Version="1.4.2" />
-    <PackageReference Include="Hangfire.Console.Extensions.Serilog" Version="1.0.2" />
-    <PackageReference Include="FastEndpoints" Version="5.15.0" />
+    <PackageReference Include="Hangfire.Console" Version="1.4.3" />
+    <PackageReference Include="Hangfire.Console.Extensions.Serilog" Version="2.0.0" />
+    <PackageReference Include="FastEndpoints" Version="5.22.0" />
   </ItemGroup>
   
   <ItemGroup>
@@ -25,23 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <_ContentIncludedByDefault Remove="Features\Accounts\Login.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\Edit.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\Forgot.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\Reset.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Edit.mail.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Edit.turbo.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Forgot.mail.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Forgot.turbo.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Reset.mail.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Passwords\_Reset.turbo.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\Register.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\_Register.turbo.cshtml" />
-    <_ContentIncludedByDefault Remove="Features\Accounts\_Registered.mail.cshtml" />
+    <Folder Include="Framework\Helpers\" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Include="Features\Orders\List.cshtml" />
-  </ItemGroup>
-  
 </Project>
